qcedev: add support for key index
Use key indexing to support multiple keys per pipe. This will
free up other pipes and allow for multiple offload DRM sessions.
Now that we have this feature, we will only be using one pipe for
all use cases.

Key index requires a new ioctl for key-index. This extends the
previous "offload" ioctl with key index and also adds parameters
that are likely to be used in the future.

Change-Id: Id8427f71e785895847bb926615d16a825402f30b
Signed-off-by: Daniel Perez-Zoghbi <quic_dperezzo@quicinc.com>