qcedev: add recovery request to prevent SID fault
When running in key-index mode, an SID fault occurs on the next
crypto operation whenever qce_manage_timout is called. This change
adds a "recovery request" which resets the engine to avoid this.

Information on reproducing the issue:
1. Create a request which has an expected error from GPCE,
   a DRM key must be used in key-index mode to pull in the
   SID from this request.
2. The engine encounters the error, and qcedev times out calling
   qce_manage_timeout.
3. The engine is reset into a proper state STATUS=0x02000004.
4. Trigger another operation in GPCE using different SIDs than
   the operation in step 1, and DO NOT use a DRM key.
5. We get an SMMU context fault and crash soon after.

This change fixes this issue by inserting a "recovery request"
in between step 3 and 4, which uses a DRM key with kernel SIDs
and is expected to complete without errors.

Change-Id: I71684d94b574df35c47a16913491e58dd1f251dd
Signed-off-by: Daniel Perez-Zoghbi <quic_dperezzo@quicinc.com>