tz_log: Add support for query log status scm call
qcom_scm_query_log_status wraps encryption and logging status for tz and
qsee logs. Add support for this scm call to update behavior on external
device as currently an intentional crash is observed which is handled
gracefully with this scm call.

In case of old QTEE firmware, use qcom_scm_query_log_encrypted to
fallback to old behavior.

Change-Id: I96af26f3947f6a776562aea8d35ca4378b298fa1
Signed-off-by: Kuldeep Singh <quic_kuldsing@quicinc.com>