securemsm-kernel: Update qseecom_proxy dependency for canoe
Enabling qseecom_proxy config on canoe requires the qseecom_proxy
module dependency to be explicitly added to canoe sec-kernel build
as now the qseecom_proxy module has migrated to soc-repo.

This change is needed for enabling CONFIG_QSEECOM_PROXY for canoe.
Without this change if we try to enable the config for qseecom_proxy
- modpost error is observed for provide_qseecom_kernel_fun_ops
API as this is exported from qseecom_proxy module which has now
migrated to soc-repo.

Change-Id: I05e6c1fec158b92cea88e80c6a08be1e4c997b75
Signed-off-by: Divisha Bisht <quic_divibish@quicinc.com>