securemsm-kernel: crypto-qti: add support for software keys
When using a copy operation in the offload use case, a DRM key index
is provided because a copy-op is an encrypt operation with the
encrypted segment size = 0. This means that a key needs to be provided
with the proper key policy. None of this is necessary. Instead, using
a software key in the proper pipe will allow for a copy operation.

This change adds support for a key to be set in software.

Change-Id: I4d93868e2294e60b18d4ae4e2088e83abe620ce6
Signed-off-by: Daniel Perez-Zoghbi <quic_dperezzo@quicinc.com>