qce: add an error cb for recovery request too
Recovery request currently does not have an error callback
defined. Although, it is unlikely to happen, it can cause a kernel
panic if an error irq is triggered for the recovery request.

Adding an error callback so this scenario is not fatal.

Change-Id: Ida711b5efbe2d0706ceb32942647a3e0848dcbb3
Signed-off-by: Gaurav Kashyap <quic_gaurkash@quicinc.com>
(cherry picked from commit 2b61d22a6533c7339ef5e6e7101a419d3cb11290)