securemsm-kernel: Update qseecom_proxy dependency for parrot
Enabling qseecom_proxy config on parrot requires the qseecom_proxy
module dependency to be explicitly added to parrot sec-kernel build
as now the qseecom_proxy module has migrated to soc-repo.

This change is needed for enabling CONFIG_QSEECOM_PROXY for parrot.
Without this change if we try to enable the config for qseecom_proxy
- modpost error is observed for provide_qseecom_kernel_fun_ops
API as this is exported from qseecom_proxy module which has now
migrated to soc-repo.

Change-Id: I51a1672baf1da9d87fa82164b52e836e2c6d59ec
Signed-off-by: Sangam Swamy <sangaman@qti.qualcomm.com>