securemsm-kernel: qce: Add kernel version check for timer container macro
Update timer callback functions in qce50.c and
qcrypto.c to use the appropriate macro for retrieving
the container structure from the timer, based on the
kernel version.

For kernels older than 6.16.0, the legacy from_timer()
macro is used. For kernel 6.16.0 and newer, the new
timer_container_of() macro is used. This ensures
compatibility with both older and newer Linux kernel
versions and prevents build errors due to macro
changes in kernel headers.

Change-Id: I2aa017f1ff0e7ec43a04ec488e10cadf33673dcf
Signed-off-by: Udit Tiwari <utiwari@qti.qualcomm.com>