securemsm-kernel: Update qseecom_proxy dependency for qcs610
Enabling qseecom_proxy config on qcs610 requires the qseecom_proxy
module dependency to be explicitly added to qcs610 sec-kernel build
as now the qseecom_proxy module has migrated to soc-repo.

This change is needed for enabling CONFIG_QSEECOM_PROXY for qcs610.
Without this change if we try to enable the config for qseecom_proxy
- modpost error is observed for provide_qseecom_kernel_fun_ops
API as this is exported from qseecom_proxy module which has now
migrated to soc-repo.

Change-Id: I9fb6c2c93b61d12c53ecd8f48f6bd1712ef1b8a1
Signed-off-by: Madhukar Sharma <madhukar@qti.qualcomm.com>